Include undefined optional schema properties as None in dto dictionary
Previously optional properties were not set on the dto dictionary if not
given in the request. This leads to dictionaries with varying keys of
the same type. Since these dictionaries represent transfer objects,
their varying keys are analogue to varying attributes of transfer
objects.

Objects have a set of attributes, that are predefined. We may construct
objects with some or all attributes set to None, if these attributes
have no value (yet), but they should always be present.

The same principle should be applied to these dto dicitionaries.